refactor(web): move session lifecycle behind SessionDriver implementations
Introduces the SessionDriver interface (create/delete/archive/rename/
ensureReady/cancel) with a static registry, and moves the flavor-specific
bodies out of SessionModule/MessageModule verbatim:

- LocalSessionDriver: agentAPI-backed creation, backend delete/archive/
  rename, state-machine cancellation.
- DispatchSessionDriver: observer-projection creation (session id + job id
  minting, observer registration), tombstoned dismissal for delete/archive,
  local-only rename, dispatchApi cancel.

Shared orchestration (workspace/title resolution, creation dedupe, cascade
computation, transient handling) stays in SessionModule and now calls the
driver at the divergence point. Model-resolution helpers and turn
cancellation choreography move to leaf utils so drivers depend downward
only; getModelMaxTokens/markCurrentTurnItemsAsCancelled re-export from
their old homes.
