+++ src/web-ui/src/flow_chat/utils/turnCancellation.ts
@@ -1,0 +1,46 @@
+/**
+ * Store choreography for marking a session's in-flight turn as cancelled.
+ *
+ * Leaf module shared by the local session driver and flow-chat-manager.
+ */
+
+import type { FlowChatContext } from '../services/flow-chat-manager/types';
+
+export function markCurrentTurnItemsAsCancelled(
+  context: FlowChatContext,
+  sessionId: string
+): void {
+  const state = context.flowChatStore.getState();
+  const session = state.sessions.get(sessionId);
+  if (!session) return;
+
+  const lastDialogTurn = session.dialogTurns[session.dialogTurns.length - 1];
+  if (!lastDialogTurn) return;
+
+  if (lastDialogTurn.status === 'completed' || lastDialogTurn.status === 'cancelled') {
+    return;
+  }
+
+  lastDialogTurn.modelRounds.forEach(round => {
+    round.items.forEach(item => {
+      if (item.status === 'completed' || item.status === 'cancelled' || item.status === 'error') {
+        return;
+      }
+
+      context.flowChatStore.updateModelRoundItem(sessionId, lastDialogTurn.id, item.id, {
+        status: 'cancelled',
+        ...(item.type === 'text' && { isStreaming: false }),
+        ...(item.type === 'tool' && {
+          isParamsStreaming: false,
+          endTime: Date.now()
+        })
+      } as any);
+    });
+  });
+
+  context.flowChatStore.updateDialogTurn(sessionId, lastDialogTurn.id, turn => ({
+    ...turn,
+    status: 'cancelled',
+    endTime: Date.now()
+  }));
+}
